+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Added support for batch processing of multiple audio files
 * Added --verbose and muted most of the messages by default
 * Complete rewrite for better internal architecture
+* Speech speed is now configurable
 
 v1.0.0 (2018-10-01)
 -------------------